+++ problem2/problem2.cpp
@@ -5,7 +5,8 @@
 
 using namespace std;
 
-
+const double valeurMaximaleTaux = 100.0;
+const double valeurMinimale = 0.0;
 
 double entreeValide(string message, double min, double max) {
 
@@ -22,28 +23,24 @@ double entreeValide(string message, double min, double max) {
 
 int main(int argc, char* argv[]) {
 	
-
-	double pret = entreeValide("montant du pret: ", 0, INFINITY);
-	double montant = entreeValide("montant remboursee chaque mois: ",0, pret);
-	double interetAnnuel = entreeValide("taux interets annuel: " ,0, 100);
+	double pret = entreeValide("montant du pret: ", valeurMinimale, INFINITY);
+	double montant = entreeValide("montant remboursee chaque mois: ", valeurMinimale, pret);
+	double interetAnnuel = entreeValide("taux interets annuel: " ,valeurMinimale, valeurMaximaleTaux);
 	
-	double interetMensuel = interetAnnuel / 12;
+	double interetMensuel = interetAnnuel / 12.0;
 
 	int nMonth=0;
-	double sumInteret = 0;
+	double sommeInteret = 0;
 
 	while (pret>0) {
 		nMonth++;
-		sumInteret += pret * interetMensuel;           // mettre dans une variable ????
+		sommeInteret += pret * interetMensuel;
 		pret = pret * interetMensuel;
 		pret -= montant;
-		
 	}
 	string out = "la dette sera rembourser apres "+to_string(nMonth)+" mois\n" +
-				"la somme des interets est: " + to_string(sumInteret)+"$";
-	
+				"la somme des interets est: " + to_string(sommeInteret)+"$";
 	cout << out;
-	
 
 }
